+++ src/GammonX/GammonX.Server/Repository/ApiGatewayClient.cs
@@ -4,12 +4,15 @@ using GammonX.Models.Helpers;
 
 using System.Text.Json;
 
+using MatchType = GammonX.Models.Enums.MatchType;
+
 namespace GammonX.Server.Repository
 {
     // <inheritdoc />
     public class ApiGatewayClient : IRepositoryClient
     {
         private readonly HttpClient _client;
+
         private static readonly JsonSerializerOptions JsonOptions = new()
         {
             PropertyNameCaseInsensitive = true,
@@ -25,22 +28,22 @@ namespace GammonX.Server.Repository
         }
 
         // <inheritdoc />
-        public async Task<PlayerRatingResponseContract?> GetRatingAsync(Guid playerId, MatchVariant variant, CancellationToken cancellationToken)
+        public async Task<PlayerRatingResponseContract?> GetRatingAsync(Guid playerId, MatchVariant variant, MatchType type, CancellationToken cancellationToken)
         {
             try
             {
-                var url = $"players/{playerId}/rating/{variant}";
+                var url = $"players/{playerId}/rating/{variant}/{type}";
                 using var response = await _client.GetAsync(url, cancellationToken);
                 response.EnsureSuccessStatusCode();
                 var responseJson = await response.Content.ReadAsStringAsync(cancellationToken);
                 var rating = JsonSerializer.Deserialize<PlayerRatingResponseContract>(responseJson, JsonOptions);
                 return rating;
             }
-            catch (Exception ex) 
+            catch (Exception ex)
             {
-                Serilog.Log.Error(ex, "An error occurred while requesting rating for '{PlayerId}' and variant '{Variant}'", playerId, variant);
+                Serilog.Log.Error(ex, "An error occurred while requesting rating for '{PlayerId}', variant '{Variant}', and type '{Type}'", playerId, variant, type);
                 return null;
-            } 
+            }
         }
 
         // <inheritdoc />